+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 - Add missing unit test for `Abs` in `EvalExpr`.
 - Fix error in unit test for `Abs` in `EvalExpr`.
 - Add missing unit test for `Sgn` in `EvalExpr`.
+- Add missing unit test for `Int` in `EvalExpr`.
 
 ## [Unreleased] - 2026-02-14
 ### Changed